Move up space token to up org token
